refactor(setup): enforce typescript baseline for service
