--- src/infra/sources/in-memory-source-repository.js
@@ -1,13 +1,0 @@
-'use strict';
-
-function createInMemorySourceRepository(seed = []) {
-  return {
-    async listEligibleSourceIds() {
-      return seed.map((source) => source.sourceId);
-    }
-  };
-}
-
-module.exports = {
-  createInMemorySourceRepository
-};
