--- src/shared/time/now-iso.js
@@ -1,9 +1,0 @@
-'use strict';
-
-function nowIso(clock = () => new Date()) {
-  return clock().toISOString();
-}
-
-module.exports = {
-  nowIso
-};
