+++ src/shared/time/now-iso.ts
@@ -1,0 +1,5 @@
+export type Clock = () => Date;
+
+export function nowIso(clock: Clock = () => new Date()): string {
+  return clock().toISOString();
+}
